refactor(tier3): split common/common.h; clean includes and warnings
Extract linalg type aliases into linalg/linalg_types.h and XML
serialization templates into common/serialization.h. Lean common/common.h
retains only the STL includes, Observation/StateIndex typedefs, and the
math/constants.h include. Update all consumers (hmm.h, calculator.h,
trainer.h, segmental_kmeans_trainer.h, numerical_stability.h, IO headers,
src/common/common.cpp, tests/test_common.cpp) to include the appropriate
new headers directly.

Remove redundant common/common.h includes from test_pareto_distribution.cpp
and test_rayleigh_distribution.cpp; both already get Observation
transitively through their distribution header.

Resolve all clean-build warnings:
- benchmarks/src/simple_benchmark.cpp: cast nodiscard getProbability to void
- benchmarks/src/libhmm_vs_hmmlib_benchmark.cpp: discard unused HMMLib
  viterbi return value
- benchmarks/src/classic_problems_benchmark.cpp: remove unused local and
  private field ClassicHMMProblems problems
- benchmarks/src/threeway_benchmark.cpp: discard unused HMMLib viterbi
  return value
- tools/fb_contour_sweep.cpp, tools/hotspot_breakdown.cpp: mark
  FB_MAX_REDUCE_FORCE_PAIRWISE_MAX_STATES [[maybe_unused]] (only active
  under experimental preprocessor flags)
