+++ include/libhmm/math/numerical_stability.h
@@ -5,7 +5,7 @@
 #include <stdexcept>
 #include <string>
 #include <vector>
-#include "libhmm/common/common.h"
+#include "libhmm/linalg/linalg_types.h"
 
 namespace libhmm {
 namespace numerical {
